Move CapturingChecksPublisher to util package.
