+++ docs/consumers-guide.md
@@ -65,7 +65,7 @@ it is recommended that you look for this name and set it over your default check
 ## Integration Testing
 
 An implementation of `ChecksPublisher` that captures all published `ChecksDetails` is provided
-in the `test` classifier, as `io.jenkins.plugins.checks.api.test.CapturingChecksPublisher`.
+in the `test` classifier, as `io.jenkins.plugins.checks.util.CapturingChecksPublisher`.
 
 Adding the factory for this publisher as a `TestExtension` will allow inspection of published checks after running a job
 on a `JenkinsRule`.
