+++ src/test/java/io/jenkins/plugins/checks/steps/PublishChecksStepITest.java
@@ -4,7 +4,7 @@ import io.jenkins.plugins.checks.api.ChecksConclusion;
 import io.jenkins.plugins.checks.api.ChecksDetails;
 import io.jenkins.plugins.checks.api.ChecksOutput;
 import io.jenkins.plugins.checks.api.ChecksStatus;
-import io.jenkins.plugins.checks.api.test.CapturingChecksPublisher;
+import io.jenkins.plugins.checks.util.CapturingChecksPublisher;
 import io.jenkins.plugins.util.IntegrationTestWithJenkinsPerTest;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.junit.Test;
